Squashed 'seims/src/taudem/' changes from 0678025dc..74ddcd84e
74ddcd84e Merge pull request #4 from lreis2415/dev
0ce00ccb2 Merge branch 'dev' of github.com:lreis2415/TauDEM_ext into dev
7d70b26b6 Update Find OMP according to CCGL
0f4d0dd6f Merge pull request #3 from lreis2415/dev
d2d2f153c Update Actions badge [ci skip]
c9b1eaadc Add Github Actions on Win,Linux,Mac
6ec626cae Build docker images
40f8c62ef Add data and ignore data/result* dirs
30b688823 merge master branch
c6377097d Merge remote branch 'master'
c097dd775 Merge commit '85e98a503ec9db8b96f15db2b851bef45dd32bf7'
85e98a503 Squashed 'cmake/' changes from 48992f8..07734aa
a090bf055 Update CMake setting: install to bin directory
7bd55e4ea Eliminate compiler's warning
da66f6fbc Update to latest Develop branch of TauDEM
12be70e09 Eliminate comping errors and warnings
563835351 Update CMakeLists based on latest cmake subtree
e519e2c61 Merge commit 'd6e8aa9ca01b751cbb6b506c51755a84eca8ece6'
d6e8aa9ca Squashed 'cmake/' changes from 48992f8..07734aa

git-subtree-dir: seims/src/taudem
git-subtree-split: 74ddcd84e19e7a8c9a8eb505529ad68d9a712717

+++ src_ext/MultiFlowDirMaxDown.cpp
@@ -154,7 +154,7 @@ int flowdirection_mfd_md(char* dem, char* fdir, char* fportion,
         // write flow fractions data into separated raster files
         for (lyr = 1; lyr <= 8; lyr++) {
             char ffracfile[MAXLN];
-            std::string intstr = std::to_string(lyr);
+            std::string intstr = std::to_string((long long)lyr);
             intstr.insert(0, "_");
             nameadd(ffracfile, fportion, intstr.c_str());
             tiffIO ffractTIFF(ffracfile, FLOAT_TYPE, static_cast<double>(DEFAULTNODATA), srcf);
